requirements.txt
flask
gunicorn
opencensus-ext-azure
opencensus-ext-flask
